Feat(frontend): use child routes for singular norm
As for caselaw, we have a group of routes that are available to interact
with one specific norm resource. If the user has selected a norm, he
gets to a view with a sidebar navigation on the left, and some content
on the right. The content is always related to this loaded norm and can
be controlled with the navigation on the right.
This is a typical use-case for child routes with a child router view.
Thereby the "root" route of the norm includes the whole setup. This
includes the navigation, layout and loading the data of the norm. If
then has a router view in its layout where all child routes get rendered
into. Each of these child routes gets the norm data passed as property
and must not load the data by itself.
This also removes the need for a wrapper component for the layout as it
is done for caselaw right now. This "new" approach is much more simple
to use and more powerful, using native features of the Vue router.