Feat(frontend): add norm frame inputs route
This route includes the input fields to edit the frame of a norm. So far
this is only a single field.
Furthermore this does not yet work functional wise. Changing the title
will do nothing so far. This first requires some refactoring to allow
using the save button and auto-save functionality.