Feat(frontend): add save to remote composable
For case-law, we have the concept of saving changes to the remote
backend server with a button and in an automated interval. Additionally
there is some logic to display the last save date as well as error and
in progress states.
Unfortunately this logic is not re-usable and heavily tight and spread
over multiple case-law specific components. This new composable is
an attempt for a first step to make this re-usable. The next step will
be the visual component with the button and the state visualization.