Fix(frontend): fix reloading of norm frame route
When a norm was selected and the frame opened for editing and then the
user reloads the page, there was an error that the norm is undefined.
The underlying issue was simply that the component managing the child
route view has to first finish loading the norm, before it can display
the child routes. Because these depend on the norm to be loaded already.
It is not clear why this doesn't happen during normal navigation. Though
it is an issue in general that should be fixed.

The removal of the `console.log` statement is unrelated to the above
changes.