Test(frontend): fix router issues for navbar
The `NavbarSide` component is "heavily" integrating with the
`vue-router` plugin. So far it is only the `<router-link>`, but we know
already that with the next planned features we also need the composables
like `useRoute`.
Mocking the complete router with its composables and also stub the link
component is quite hard. Or rather we still have a remaining issue of
some Vue warnings during the tests where we don't understand where they
are coming from.
Therefore it is currently the strategy to do some kind of integration
tests for this component with a real router instance. As a side-effect,
this requires us to define all the routes that the navigation component
uses. That is because the link component of the router does some
"special" checks. This means while doing the tests we do not only need
to define the menu items, but also the related routes. To makes this
more simple for the test cases themselves, some utility functions have
been added for this test suite that automatically creates the routes
according to the menu items. Additionally it also adds the required root
route.
While this turned out a bit verbose, it is very simple to use, the tests
themselves are good to read and we do some good integration testing
here.