Feat(frontend): align navigation naming convention
The naming of the different menu items of the `NavbarSide` component
wasn't really strong. There was always this conflict if it might become
recursive, what is the same between different kinds of menu items and
more.
After a conversation with the Design team, we aligned the naming with
their terminology. So we now clearly differ between a level one and
a level two menu item.
Furthermore the has the name of a menu item changed to a label. This
also helps to prevent the confusion with the name of the route for
a menu item. Moreover it aligns with the go back label property.