Refactor(frontend): improve navigation setup
The `NavbarSide` component has become a little ugly with the two
different levels of menu items and the special classes for different
situations. And we already know that more features are coming which
makes this even worse.
Therefor this is an approach to organize the code a little better and
extract the logic from the template to the setup function. This results
into a much simpler template, which is always hard to read with too much
JavaScript code inside. Furthermore it allows to be more expressive with
proper variable and function names in the setup function. So it is now
clear that the last sibling of a level one's level two items gets
a special margin.
What is important to keep in mind is that the `NavbarSide` component
remains mounted while the child routes in its adjacent route view
change. This puts extra requirements on the reactivity of this
component. This also means it is not possible to calls functions from
the template `v-for` loops as these would be only re-evaluated if the
menu items change. But this will no more be the case as we will have
side-effects by other reactive factors like the current active route
etc. Therefore it has been decided to use a computed value that enhances
the menu items. This works well because this is the exact same hierarchy
that the template goes through and accessing "additional"/enriched
properties is straight forward this way.
To keep the code readable, we use separated functions to generate
specific enhancement properties which are then used by the computed
value that puts it all together. Thereby the callback of the computed
value has only a single responsibility and remains somehow readable.