Fix(frontend): identify active navigation item
According to the design, the `NavbarSide` component should highlight the
currently active menu item according to the active route.
So far we have been using the `vue-router` native functionality for
this. But this doesn't work out properly for us. There is only the
option to match the whole route (including path, query and hash) or only
the path. But we need to match the path and the hash, but ignore the
query. Furthermore it is better for us to match the route name if
possible to get more exact results (e.g. for index routes).
Therefore a custom implementation to identify an active menu item has
been implemented. While it is a quite versatile solution, it makes some
assumptions to simplify it here and there.
Also the code is a little verbose, it tries to make everything easy to
understand and builds on top of the earlier introduced reactive menu
item enhancement.