Fix(frontend): use https for URL matching
To parse URL style route locations in the `NavbarSide` component, we use
the `URL` class. While we actually only care about everything after the
protocol and host, we still have to provide them (else the constructor
fails). As workaround, an obvious placeholder has been defined. That
placeholder was using the insecure HTTP protocol.
Our quality gate is now complaining that HTTPS is recommended instead.
While it doesn't matter at all for this use-case here, it still got
changed now. It is a placeholder anyway and solving the issue like this
is "better" than ignoring it in the quality gate.