Feat(frontend): open and close navigation sections
According to design, the `NavbarSide` should automatically open and
close the different sections. This means that the level two items of all
level one items are hidden per default. Only the level two items of an
active level one item or with an active sibling are open and visible. If
the user switches to another "section" the former sections closes
automatically.

While the two functions to determine the highlighting of any menu item
and if a level one item is expanded look almost like copy-paste, it is
definitely not easy to merge both. Thereby it is much cleaner to keep
them separated.

This also caused a change on an end-to-end tests because it requires to
open the navigation section first, before the level two items are
visible and clickable for the user.