Feat(norms): slim down list norms use-case data
The list norms use-case service one purpose: listing norms in the
documentation front-end. It should only provide the necessary data of
each norm for a user to pick one and the GUID that enables to load this
specific norm.
This has been now pinned down in the use-case itself too. Therefore we
do no more communicate the whole norm entity but have a new, slim down,
data transfer object for this application interface. The HTTP controller
now only needs to take care of the JSON marshalling and the pagination
data structure.

So we added another step to a full-mapping strategy, but save ourselves
some future work. The list norms use case now only changes when it is
required to.

Because of this change, the response schema and the data transfer
objects between listing and loading a norm has changed. Due to that, the
HTTP controller has been split too. This also makes the tests separated
and very easy to follow and extend. They also have become a bit more
accurate to test for the different data models.

Finally this change also been documented in the API specification and
caused some marginal changes in the frontend too. But nothing surprising
here.