Make inputs lazy to only update header on blur
RISDEV-609

v-model.lazy reflects changes only after @changed event or after focus is lost in case of input fields, therefore the tests regarding the input fields are all using the change event instead of the user type event. Inly the dropdown can't be lazy, because we need to check for each user input to filter the dropdown items. But as the model value in the dropdown input component does not emit on value change, but only on select, the dropdown does not need to be lazy.