Docs(norms): update list norms API specification
Recently we changed the use-case (and thereby also the API) for the
listing of norms. Unfortunately we made the mistake to implement first
and specify second (now). This is syncing both again.

Note that this also renames some operation identifiers to by more in
sync with the use-case naming and avoid possible confusion.