Docs(norms): fix schema issues for norms
The OpenAPI linter was throwing some issues which pointed out that the
`Article` schema was actually not used. This was a schematic issue
related to our concept of having multiple schemas with and without
GUIDs. This got fixed now and works properly according to the API
rendering preview.
While doing so, the sorting of the schema merging has changed a little
which makes the GUIDs to appear at the top in the rendering.
Moreover some redundant examples could be removed which makes the YAML
file easier/smaller, but maintain the same quality for the final
rendering.