Chore(backend): unify norm input port mapping
Multiple changes related to the mapping between the input ports have
happened.

First of all, the de- and encoding of the `LocalDate`s and `UUID`s has
moved to the input ports. The application should not deal with these
concerns. Especially as different input ports could use different
encodings for communication.
In relation to this, there was an awkward issue with some assertions to
compare a `LocalData` with `null`. Thereby the diff is much bigger due
to the switch to a different assertion method.

Moreover the response schema data classes of the input controllers have
become unified. All of them have now private constructor and have
a companion function to create an instance with the name
`fromUseCaseData`.
Similar for request schema data classes, they have a method called
`toUseCaseData`.

Some shared en- plus decoding methods have been put into a new file and
get re-used between the different input ports of the REST API.