Chore(backend): unify norm type references
We have multiple data classes for commands and other data for the input
ports. As they must strongly relate to their specific use-case, they are
a "child" of their use-case class. So instead of having `MyUseCase` and
`MyUseCaseCommand`, the latter become `MyUseCase.Command`.
Now it happened that sometimes these children got imported directly and
used with their short scope names (e.g. just `Command`). It is tempting
because it is much shorter to use.
Firs of all there was a mixed usage which is not nice. So a unified
approach has be chosen for the moment (might change). The argument to
use the full scope syntax (e.g. `MyUseCase.Command`) is that some of
these classes have very generic names like `Command`. It can easily
happen that a class uses multiple interfaces that both have this scope
class. At least then it becomes necessary to differ between both and the
short syntax is no more possible.