Refactor(backend): split edit norm frame command
The command for the use-case to edit a norms frame consists logically of
two parts: the data to identify the norm and the frame properties to
edit. Currently the command is just a flat object that makes it hard to
grasp this logic and possibly even harder to extend. Applying the above
logic as code makes this more clean.
A "proof" of this is the better usage within the API controller. Though
the application interface should not be dictated by the adapters, it is
good hint. For the controller the identification information is
separated from the properties. Having the new structures allows for
a better data mapping and decoding with the common `toUseCaseData`
method. The route function itself also becomes much better to read.

Note that for the mapping of request data to use-case data, the methods
have been adapted to make use of the keyword `this` to address
properties of the class. While this is optional and not necessary in
Kotlin, it helps to read the function better. As most of the time all
class property names are equal, it become a little "messy". Clarifying
that one is the parameter name and the other a property helps.
The same goes already for the opposite direction where it is clear what
is the parameter name and what the property of the input (here it is not
optional).