Tests(frontend): skip edit norm frame e2e test
We have some issues with the recent changes and the end-to-end test we
have to sort out. We struggle to combine HTML tag locators with exact
text matches. So because we have multiple headings with similar text
content, the tests struggle to uniquely locate some stuff.

But right now the test has become kinda obsolete anyway because the
use-case has changed dramatically. So we agreed that it is fine to skip
the test for now as we will re-work it anyway.