Feat(frontend): add state management for norm
When a user selects a norm, he enters a new space of the application.
This space has to load the norm and all components/views act on this
norm.
To simplify the data flow and avoid to much boilerplate for top-down
communication, a store for state management has been added. The store
holds the currently loaded norm. If the user leaves the space, it will
be unloaded again. The store then provides some functions to act on the
currently loaded norm.
This for examples allows to easily change some norm data in one view and
make it instantly available in other views too. It also simplifies the
updating of the data to a remote by having in one central place.

Please note that this store is not yet used by any component. This will
follow soon, keeping this commit focused.