Feat(frontend): make use of loaded norm store
After the new state management got added, the store gets used in the
components now.
This finally simplified the parent-child communication, which was the
cause for introducing a store. Thereby some properties and events could
be removed.
Note that this also affected the `SaveButton` component
which was involved into that process too. In this case it is even better
because the button has become agnostic again at not related to the norms
context anymore.

Because we still have to map between the `Norm` data model and the input
models, we keep the computed value for the norm frame. Though it is
quite verbose. We will keep an eye on that as the whole frame thing is
a moving topic anyway.

Moreover, because the `useStoreToRemote` composable currently expects
a service response as return type, this was added to the loaded norm
store update function. Unfortunately we currently can't handle the case
of a "skipped" save/update. Therefore we just return a "successful"
response in case it was skipped.
We definitely need to iterate over this when the picture become more
clear how this should be orchestrated in future. Also related to the
topic of not saving if nothing has changed.