Fix(frontend): add proper norm date encoding
This extends/"reworks" the encoding of the norm frame data send to the
API. A bug was found where the send date format does not match the
exceptions. Thereby it was not possible to save a norm anymore once
a date was selected.
The issue is that currently the date input components represents date in
this format: `2022-11-14T23:00:00.000Z`. The expected format by the API
is: `2022-11-14`. Therefore the encoding in the norms service had to
adapt.
We should consider to use a proper date type for the input model. This
would also enforce a proper encoding, as a rich type could not be
serialized automatically as it happens for the plain string.