Chore: update a bunch of missed talisman checksums
Because we recently had to collect a lot of commits locally and the
Talisman check does only run on push, a bunch of checksums have become
outdated. As it would a big struggle to interactively rebase each
commit, this standalone one was made.
We need to find a better solution for this hell.