Add TupleInput element
RISDEV-487

In order to display nested, toggable inputs, we add a tuple element, which takes a parent and child field config. The model values for this input type accept an object of parent und child model values, which needs some mapping in DocumentUnitCoreData. There we need to map the CoreData, to structured core data, which our components can work with.