Tests(frontend): prevent console usage
We agreed to make the frontend "unit" a-like tests kinda silent. Should
mean the output includes only the information of which test run
successfully and the result of failed tests.
Any kind of warning or usage of other logging to the console is
forbidden. This is so that warnings do not get ignored and tests are
written properly. For example when something was not mocked properly or
similar. Furthermore is the usage of console.log not a good practice.
In the past we had a mass of additional output and the team got
desensitized to catch warnings as the additional output was just
ignored as unimportant.
With the recent effort we made it to fix all warnings in the tests. With
this commit now, it helps us to keep it like that by making the whole
test suite fail if any test uses the console. So the developer must
first fix his test again before he can commit his changes.

There were some considerations to write this functionality quickly
ourselves. But it is not just about 10 lines of code and there is a nice
dependency we can add. I have already some "experience" with it and is
easy to use and works reliably.