Add useTransformTupleData composable
RISDEV-487
This composable takes two keys to transform flat data to structured data (into a tupel), which the frontend then could display. It returns a computed property with getter and setter, so when the modelValue changes, the data could be transferred back to flat data for the backend. The new tupleKey is autogenerated after the pattern parentKeyAndChildKey