Chore(tests): add playwright linting plugin
This adds a plugin for the ESLint linter that adds some more rules for
the usage with Playwright.
Due to the extension, also a couple of other linting rules run on the
end-to-end tests which have not before. But all of them where
auto-fixable.

Two of the issues had to be ignored manually as we definitely need them.
We tried to verify in the Playwright runner is independent of the
operation system to send the Meta or Control key, but it is not.