Feat(norms): add output port to convert norms
Add a new port for the application that allows to provide functionality
to convert norms to LegalDocML.
For the moment, this is a pure structural change. The only output
adapter implementing this functionality still just returns a static XML
string. It will be the goal of the following commits to add that
logic.