Refactor(norms): adopt more ArchUnit concepts
This heavily refactors the previously existing ArchUnit predicates and
conditions. It become all much more re-usable and prepared to move into
a different module if necessary. Moreover it adopts much more the
concepts of ArchUnit is structures itself.

All conditions are always based on a predicate and mostly get just
straight converted from the predicate. Though some conditions are still
a class to allow for more detailed error events with better output for
the user if something fails. So for example if a predicates determines
an expected value dynamically based on inputs, it is super helpful to be
able to read that in the error message too.
Everything must be in first place a predicate to make it re-usable from
the beginning. Also we only need it as condition now, having it as
predicate allows to use it for both cases easily.

There is still a "big" issue with naming collisions between predicates
and conditions sometimes. There are a lot of options to wrap them with
`are`, `be`, `is`, `have`, `not`, ... etc. But sometimes it just hard.
Unfortunately is is not easily possible to extend the `ClassesThat` and
`ClassesShould` to avoid the collisions and have even better
streamability.