Refactor(norms): adapt ArchUnit that and should
Make sure to use almost everywhere parameter notations for `that()` and
`should()` in the ArchUnit tests. Our own predicates and conditions
already differ because they simply can't be streamlined. This makes it
more homogeneous. And partially even saves a few lines of code. At least
within a test case, as this requires more imports.
But many tests now simply read a that-should-check in single lines which
is quite neat.