Refactor(norms): use more functions for ArchUnit
This makes all custom predicates and conditions a function. Before we
have used values for some rules which don't have a parameter. While this
quite neat, it was not very homogeneous. And in contrast are all ArchUnit
library predicates and conditions functions. Having that changed, the
tests look better now. Especially with some highlighting effects of
code.