Test(norms): add coding rules test
So far this only includes a few useful rules from the ArchUnit library.
But it is intended to add more rules here of some parts we have agreed
on as a team.

Furthermore this starts to split the classes into source-, test-, and
all classes to allow scoping rules better.