Test(norms): verify domain packages
A simple test that ensure only the expected packages exist in the domain
based on what we decided. This should prevent us from simply putting
a class "somewhere". If this test fails, the developer should rethink
his choice or communicate with the team if we need to have a new package
and add it to the test.

This also fixes the already existing rule breach where the `value`
package was below the `entity` even though was decided to keep them
siblings.