Test(norms): ensure correct assertions are used
The team has decided to use org.assertj.core.api.Assertions everywhere
because it reads much better. This test helps us to ensure that. Also it
isn't easy to prevent any other library, we can at least exclude others
we have a dependency.
The class is imported on purpose so this gets removed once the
dependency got removed.

This fixes all rule breaches.