Fix(backend): add ignore for CVE-2022-1471
Unfortunately we strongly depend on the `org.yaml:snakeyaml` library
which has another vulnerability found. There is no old version without
a vulnerability we could go back to.
There is already a new version planned to fix this issue. Till then, we
have to ignore it and keep subscribed to get updated immediately.