Chore(hooks): run security audit pre-commit
This also runs the security audit before a commit. So far this was only
checked when all commits are already created and they get pushed.
The difference is that on pre-commit, Talisman checks the diff only. On
pre-push, Talisman checks each affected file as whole. The latter is
what the pipeline does to. This might catch very rare cases. Therefore
we keep both. The pre-push should be no issue if every pre-commit check
went fine.

This was fixed because many developer struggled with this setup. Often
this simply lead to an extra "fix talisman checksums" commit. This helps
to update checksums early at the commit that changes the files content.

Furthermore, there is a hint now that suggest to use a command that runs
Talisman in interactive mode that helps to apply checksum changes into
the runtime configuration of Talisman. This addresses another friction
point we had.
Thereby the experimental, custom script is obsolete.