Chore: rename web to frontend for alignment
This is the first commit targeting the mono repository re-structuring.
It applies the renaming of the `web` folder to be named as the system
component it implements: the `frontend`.
This requires a lot of changes also for the pipeline, documentation,
ignore patterns, etc. pp.

The background is that we actually have a mono repository, but it
doesn't like such one. The system component that inside this repository
are not identifiable as such. This means it also doesn't align with our
documentation.
Therefore we have agreed to adapt the repository structure.