Chore: new folder for backend component
This moves all files that are related to the backend component into its
own directory. Thereby we have now a repository structure where each
system component resides in its own folder. This keeps thing more tidy
and clean.
This required a lot of changed around it as the backend component was
strongly tied to the overall repository itself, especially as everything
was implicitly the "root". Now things must be more explicit and split.

The background is that we actually have a mono repository, but it
doesn't like such one. The system component that inside this repository
are not identifiable as such. This means it also doesn't align with our
documentation.
Therefore we have agreed to adapt the repository structure.