Test(norms): make xml mapping testable
This applies a few tiny changes to allow the norms XML mapping functions
to have (better) unit tests.

By reducing the level of complexity to map a full list of articles
or paragraphs, these functions only map a single one now. This comes
also handy for the tests. It is rather the job of the mapping logic by
the parent mapper to apply this and iterate other lists or similar.

Furthermore all functions start with the verb `map` now for better
readability and clean code. The Postfix of `ToDto` should be
re-considered later on.

Moreover do the mapper functions now address constructor parameters by
name. This makes it much easier to extend, remove or generally adapt the
data transfer objects. It throws errors early and makes working with
them simpler. Also commit diffs might read better.