Feat(norms): add norm markers to xml
This correctly parses the marker name attributes for norm articles and
paragraphs based on the regular expression from the LegalDocML grammar.
As here is currently a gap between the LegalDocML standard and our
domain model, we have to compensate the fact that currently our model
might miss a (correct) marker. For such a scenario we currently fall
back to the ordinal number as marker, which is the default in
LegalDocML.

As a small adjustment, also related to previous commits, the XML
converter DTOs aim to only persist of simple data types, mostly Strings
only. This is so we do not rely implicitly on `toString` functions of
certain data types. We might wanna add a code convention test for this.