Feat(norms): add correct eIds to norms XML
This uses the recently added markers to generate correct eIds for the
articles and paragraphs.
We decided to keep as many eId(parts) as possible static in the XML
template. Unfortunately it is not possible to access the parent-loop its
properties. Therefore we decided to go with the quick (but very ugly)
fix to store the article marker also into the paragraph data object to
make it accessible within its scope to generate the eIds of the
paragraph (which depend on the article eId).