Chore: move ESLint configuration from script
Within the `package.json` file, a few scripts use `eslint` and pass some
additional properties to it. As scripts in a `package.json` file are
always a bit hard to read and the shorter the better, plus this
configuration should apply always, it has been moved into the ESLint
specific configuration file.