Chore: make frontend scripts more flexible
This adapts the frontend scripts related to formatting and linting. It
allows to run them with a list of files only. Per default they work as
before and run for all files. But if an additional list of parameters is
provided, the scripts interpret them as a list of files.

While doing this, the scripts got tidied up a little. Each linting and
formatting tool we use has it's own script now with a `:check` and
`:fix` version. This is partially necessary because we use now functions
within the scrips which makes merging hard. Furthermore do the tools
vary to switch between check only and fixing mode. All of them then get
collected in the previously known `style:check` and `style:fix` scripts.
Note that the `check` requires all too succeed, while the `fix` runs
always all tools.