Chore: improve runtime of pre-commit hook
This improves the runtime of the pre-commit hook. This comes from
three changes.
First of all, it runs all frontend checks as separate commands to
parallelize them. The disadvantage is that git-hooks and script can now
diverge a little. But that is already the case anyway.
Second, the frontend command run only with the staged files now.
Third, due to the latter change, this automatically means that if
someone only works on the backend, the frontend commands are simply
skipped and not executed at all.

Additionally, to keep the noise a little down, the hooks have become
less verbose. So there is no more output of successful commands.
Moreover do the frontend commands not print out their whole script body,
but only the issues they found.