Chore: allow to skip backend commit hook
This also makes use of forwarding only the staged files to the
pre-commit hook command in Lefthook. Though is kinda hard to tell if
Spotless is accepting this list right now.
Anyway we gain the advantage that Lefthook will skip the backend command
if no backend files have been touched. Spotless is super fast anyway.
Though it can be annoying to get reports for files which are not
staged.