Chore: fix Lefthook root paths
The Lefthook configuration was broken. In result all commands that had
have a root path configured, were always skipped. Thereby no formatting
or linting was run anymore.
The issue is that Lefthooks does not work well with relative paths. We
thought that this support was added with the recent version. But we
forgot to test it again and it does not work.
To fix the issue, the `./` prefix of all all `root` properties got
removed. This has proven to work on multiple setups.