Chore: fix Lefthook command for backend
The command for Lefthook to verify the backend style was working well
with a single file as it was used for test purpose. As soon as a second
file is staged, this throws an error. The issue is that Lefthook
replaces the `{staged_files}` get replaced with every file path as an
individual string in own quotation marks with a space in-between. So
only the first file gets used by the `-DspotlessFiles` argument
assignment. The result becomes unknown.
This was solved by using a streamline editor and replacing the
whitespace characters between the file paths with commas, as Spotless
documents.