Chore: fix ESLint hook ignore issue
The issue was that if ESLint gets called with an explicit list of files
to check, it will throw warnings for files that are configured to be
ignored by ESLint. It is quite controversially discussed if this should
be fixed by ESLint or not.
As a work-around, we configure a list of patterns for Lefthooks. It then
filters the list of files it will provide to the command.
This is not so great, because this splits the configuration of the
ESLint tool. But at the same time it was basically something we kinda
always had beforehand. It is a compromise.
We subscribe to changes of ESLint and the RFC to fix the issue directly.