Chore: fix format backend pre-commit hook
When having more than 2 backend related files staged, the backend
formatting hook was failing. The problem was that the streamline editor
command to convert the list of staged file from the Lefthook format to
the format for Spotless, it was doing it only for the first two files of
the list.
Adding the `g` (global) flag, applies this substitution command for the
full list.