Chore: improve post-commit secret hook
This does two changes. First and most important, it skips the
post-commit hook when doing a rebase.
This comes with a big problem. If you do a rebase, the whole "fake call"
to Talisman is off. It is very hard to fake a pre-push hook during
a rebase. Thereby Talisman will report a mass of false positive reports
during the rebase commits. This confuses the user and is not helpful
anymore.
On the other hand, rebasing can change the content of the commits
(obviously), especially with auto-merging. Then the checksums in the
commits are actually off and wrong.
Since this is still caught by the pre-push hook, we are safe. But it
not optimal at all. Though, still a little better than before. It
basically helps for "normal" work without file merges due to rebasing.
Else it is like it always was. Maybe I'll find a solution to generate
proper pre-push arguments during a post-commit while doing a rebase.

The second change was an additional custom hook message to explain the
user what it means if Talisman fails during a post-commit and what to
actually do now.