Feat(norms): add optional query to list norms
The use-case to list norms has been extended that allows to filter the
output with a search term. This term is used in all title related
properties of a norm.
This required to extend the list of searchable fields on the related
output adapter.

So far the list norms service switches between the two output ports to
get all norms if no query was provided or to use the search with
a constructed query. This might become optimised later on.