Chore(norms): make edit norm frame port a command
To stick more strictly to our architecture conventions, also output
ports should be always be either a command or a query. Using singular
arguments for the ports with a member type is the choice to make this
explicitly stated and testable.
Note that the output ports are currently not tested for this convention
in the convention/architecture tests.