Chore(norms): make convert norm port a command
To stick more strictly to our architecture conventions, also output
ports should be always be either a command or a query. Using singular
arguments for the ports with a member type is the choice to make this
explicitly stated and testable.
Note that the output ports are currently not tested for this convention
in the convention/architecture tests.

The tests of the related service also got changed to test only what is
done by the service. As all output ports are mocked, it isn't possible
to test their behavior anyway.
Moreover we started to use `block()` for the service executions if we
don't care about the output but only the side-effects. It simplifies the
tests and states more obviously that this doesn't matter for this test
case. This will be extended soon to more tests.