Chore(norms): make search norm port a query
To stick more strictly to our architecture conventions, also output
ports should be always be either a command or a query. Using singular
arguments for the ports with a member type is the choice to make this
explicitly stated and testable.
Note that the output ports are currently not tested for this convention
in the convention/architecture tests.

This also continues to use the Reactor blocking approach in tests where
we don't are about the results.