Chore(norms): remove obsolete memory adapter
This was the very first adapter before we had a database. For some time
it was still used for simple local tests without a database.
But by now it has become completely outdated as it supports only a small
subset of output ports and there is no other output adapter that could
be used to complement it. Only the database adapter can be used by now.
Therefore this legacy piece of code has no use anymore and can be
removed.